requirements.txt
pandas
matplotlib
seaborn
sklearn
watchdog
opencv-python
wordcloud
nltk
openpyxl
streamlit-option-menu
Pillow
tensorflow
streamlit
bokeh==2.4.3
google-cloud-storage
#streamlit-nested-layout
#google-api-python-client
#google-auth-httplib2
#google-auth-oauthlib